backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
tensorflow>=2.10.0
pillow>=9.5.0
numpy>=1.23.0
opencv-python>=4.7.0
